requirements.txt
py-cord==2.3.3.dev3+gc2c45596
pyshp==2.3.1
python-dotenv==0.21.0
requests==2.28.1
Shapely==1.8.5.post1
